// frontend/src/app/openapi/model/git-instance.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { GitType } from './git-type';


export interface GitInstance { 
    type: GitType;
    name: string;
    url: string;
    api_url: string | null;
    id: number;
}
export namespace GitInstance {
}